null checks on authorize

+++ backend/src/main/java/pl/aleokaz/backend/friends/FriendsController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -26,6 +27,9 @@ public class FriendsController {
     @GetMapping("/all")
     public ResponseEntity<List<FriendDTO>> getFriends(Authentication authentication) {
         try {
+            if(authentication == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+            }
             UUID currentUserId = UUID.fromString((String) authentication.getPrincipal());
             return ResponseEntity.ok().body(friendsService.getFriends(currentUserId));
         } catch (Exception e) {
@@ -36,6 +40,9 @@ public class FriendsController {
     @PostMapping("/add")
     public ResponseEntity<ResponseMsgDto> addFriend(Authentication authentication, @RequestBody FriendCommand addFriendCommand) {
         try {
+            if(authentication == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+            }
             UUID currentUserId = UUID.fromString((String) authentication.getPrincipal());
             FriendsService.FriendStatus status = friendsService.addFriend(addFriendCommand, currentUserId);
             return ResponseEntity.ok().body(ResponseMsgDto.builder().message(status.name()).build());
@@ -47,6 +54,9 @@ public class FriendsController {
     @PostMapping("/remove")
     public ResponseEntity<ResponseMsgDto> removeFriends(Authentication authentication, @RequestBody FriendCommand removeFriendCommand) {
         try {
+            if(authentication == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+            }
             UUID currentUserId = UUID.fromString((String) authentication.getPrincipal());
             FriendsService.FriendStatus status =  friendsService.removeFriend(removeFriendCommand, currentUserId);
             return ResponseEntity.ok().body(ResponseMsgDto.builder().message(status.name()).build());
